expense-splitter-main/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
typing-extensions>=4.8.0
pymongo==4.6.1
motor==3.3.2
python-multipart==0.0.6
jinja2==3.1.2
dnspython==2.6.1
